Server renamed to bot; telegram logic from Server replaced to TelegramBot for better decomposition and work with other platforms in future
