+++ MatchmakerBot/src/com/company/Main.java
@@ -4,15 +4,14 @@ import org.telegram.telegrambots.ApiContextInitializer;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiRequestException;
 
-import java.util.Scanner;
-
 public class Main {
 
     public static void main(String[] args) {
+        Bot bot = new Bot();
         ApiContextInitializer.init();
         TelegramBotsApi telegramBotsApi = new TelegramBotsApi();
         try {
-            telegramBotsApi.registerBot(new Server());
+            telegramBotsApi.registerBot(new TelegramBot(bot));
         } catch (TelegramApiRequestException e) {
             e.printStackTrace();
         }
